Move TsuidezakeService interface to repository module.
