+++ data/remote/build.gradle.kts
@@ -32,6 +32,7 @@ android {
 dependencies {
     implementation(project(":model"))
     implementation(project(":data:auth"))
+    implementation(project(":data:repository"))
 
     implementation(Deps.Kotlin.stdlib)
     implementation(Deps.Kotlin.coroutinesCore)
