+++ data/remote/src/main/java/jp/kuaddo/tsuidezake/data/remote/RemoteDataComponent.kt
@@ -5,6 +5,7 @@ import dagger.Component
 import jp.kuaddo.tsuidezake.data.auth.AuthService
 import jp.kuaddo.tsuidezake.data.remote.internal.di.RemoteDataModule
 import jp.kuaddo.tsuidezake.data.remote.internal.di.RemoteDataScope
+import jp.kuaddo.tsuidezake.data.repository.TsuidezakeService
 
 @RemoteDataScope
 @Component(
