+++ data/remote/src/main/java/jp/kuaddo/tsuidezake/data/remote/internal/TsuidezakeServiceImpl.kt
@@ -5,17 +5,16 @@ import com.apollographql.apollo.ApolloClient
 import com.google.firebase.storage.FirebaseStorage
 import jp.kuaddo.tsuidezake.data.remote.AddSakeToTastedListMutation
 import jp.kuaddo.tsuidezake.data.remote.AddSakeToWishListMutation
-import jp.kuaddo.tsuidezake.data.remote.ApiResponse
 import jp.kuaddo.tsuidezake.data.remote.RankingsQuery
 import jp.kuaddo.tsuidezake.data.remote.RecommendedSakeQuery
 import jp.kuaddo.tsuidezake.data.remote.RemoveSakeFromTastedListMutation
 import jp.kuaddo.tsuidezake.data.remote.RemoveSakeFromWishListMutation
 import jp.kuaddo.tsuidezake.data.remote.SakeQuery
-import jp.kuaddo.tsuidezake.data.remote.TsuidezakeService
 import jp.kuaddo.tsuidezake.data.remote.WishListQuery
 import jp.kuaddo.tsuidezake.data.remote.fragment.ContentFragment
 import jp.kuaddo.tsuidezake.data.remote.fragment.SakeDetailFragment
-import jp.kuaddo.tsuidezake.data.remote.toApiResponse
+import jp.kuaddo.tsuidezake.data.repository.ApiResponse
+import jp.kuaddo.tsuidezake.data.repository.TsuidezakeService
 import jp.kuaddo.tsuidezake.model.FoodCategory
 import jp.kuaddo.tsuidezake.model.Ranking
 import jp.kuaddo.tsuidezake.model.SakeDetail
