+++ data/repository/src/main/java/jp/kuaddo/tsuidezake/data/repository/ApiResponse.kt
@@ -1,0 +1,7 @@
+package jp.kuaddo.tsuidezake.data.repository
+
+sealed class ApiResponse<out T : Any>
+
+data class SuccessResponse<T : Any>(val data: T) : ApiResponse<T>()
+
+data class ErrorResponse(val message: String) : ApiResponse<Nothing>()
